+++ build-installer.ps1
@@ -1,6 +1,6 @@
 param(
     [string]$PortableZip = "",
-    [string]$Version = "0.12.1"
+    [string]$Version = "0.12.2"
 )
 
 $ErrorActionPreference = "Stop"