+++ package.ps1
@@ -5,7 +5,7 @@ param(
     [Parameter(Mandatory = $true)]
     [string]$HeadlessCorePath,
 
-    [string]$Version = "0.12.1"
+    [string]$Version = "0.12.2"
 )
 
 $ErrorActionPreference = "Stop"