+++ release.ps1
@@ -1,5 +1,5 @@
 param(
-    [string]$Version = "0.12.1",
+    [string]$Version = "0.12.2",
 
     [string]$RuntimePath = ".\artifacts\headless-runtime",